refactor: interface+impl 구조로 리팩토링
